homework done

+++ hw09-spring-context/src/main/java/ru/otus/pro/kovaleva/model/Product.java
@@ -1,29 +1,39 @@
 package ru.otus.pro.kovaleva.model;
 
-import java.math.BigDecimal;
-
 public class Product {
-    private Long id;
+    private int id;
     private String name;
-    private int price;
+    private double price;
 
-    public Product(Long id, String name, int price) {
+    public Product(int id, String name, double price) {
         this.id = id;
         this.name = name;
         this.price = price;
     }
 
-    public Long getId() { return id; }
+    public int getId() {
+        return id;
+    }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(int id) {
+        this.id = id;
+    }
 
-    public String getName() { return name; }
+    public String getName() {
+        return name;
+    }
 
-    public void setName(String name) { this.name = name; }
+    public void setName(String name) {
+        this.name = name;
+    }
 
-    public Integer getPrice() { return price; }
+    public double getPrice() {
+        return price;
+    }
 
-    public void setPrice(Integer price) { this.price = price; }
+    public void setPrice(double price) {
+        this.price = price;
+    }
 
     @Override
     public String toString() {
